fix: Move EntityMoveProcessor to portals module, only check entities in worlds with activated portals
EntityMoveProcessor can be quite expensive and is only used for portals. Moving
it to the portals module means the performance cost is only paid for servers
that use portals. There's also no need to check for movement of all entities
when no portals are activated - instead, only check entities in worlds with
active portals.