fix: stop portals in The End generating end gateway strucutres when used
If an entity touches an end gateway block in The End (or a world like it) which
has no exit location set, the game will generate a new end gateway structure at
the same angle relative to and 1024 or so blocks away from the origin (i.e.
around the inner edge of the outer islands). Using a vane portal in an End world
therefore creates a tower of gateways that grows each time the portal is used.
This happens before Entity/PlayerTeleportEndGatewayEvents are fired, so the
existing cancelling of them doesn't prevent this.

Setting an exit location for end gateway blocks in vane portals is enough to
prevent this behaviour. Any location will do, since the teleport events being
cancelled will prevent anything going there anyway.